requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
aiohttp==3.9.1
opensearch-py==2.4.0
prometheus-client==0.19.0
google-generativeai==0.3.0
python-dotenv==1.0.0
numpy==1.24.3
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2